third_party/blink/web_tests/platform/generic/fast/dom/Document/CaretRangeFromPoint/caretRangeFromPoint-in-user-select-none-expected.txt

PASS range.startContainer is $("sample").firstChild
PASS range.startOffset is 2
PASS range.collapsed is true
PASS successfullyParsed is true

TEST COMPLETE
